+++ src/Modul3_3/Employee.java
@@ -1,18 +1,47 @@
-package Modul3_3;
+package src.Modul3_3;
 import java.util.Random;
 
 public class Employee {
 
+    String name;
+    String surname;
+    String status;
+    int id;
+
     Random random = new Random();
 
-    Employee(){
-        int id = random.nextInt(100);
+    public void setID(){
+        id = 1;
     }
 
-    Employee(String n, String s, String st){
-
+    int getID(){
+        return id;
     }
 
+    public void setIDRandom(){
+        id = random.nextInt(1);
     }
+
+    int getIDRandom() {
+        id +=1;
+        return id;
+    }
+
+    public void setName(String name){
+        this.name = name;
+    }
+    String getName(){
+        return this.name;
+    }
+
+    String getSurname(){
+        return surname;
+    }
+
+    String getStatus(){
+        return status;
+    }
+
+}
