+++ src/Modul3_3/ListEmployee.java
@@ -1,5 +1,4 @@
-package Modul3_3;
-
+package src.Modul3_3;
 import java.util.Scanner;
 
 public class ListEmployee {
@@ -11,20 +10,21 @@ public class ListEmployee {
         System.out.println("Enter your name:  ");
         String name = in.nextLine();
 
+        /*
         System.out.println("Enter your surname:  ");
         String surname = in.nextLine();
 
         System.out.println("Enter your status:  ");
         String status = in.nextLine();
 
-        Employee person1 = new Information();
-        Employee person2 = new Employee("Ivan", "Ivanov", "single");
+*/
+        Information person1 = new Information();
 
-        System.out.println(((Information) person1).getName());
-        System.out.println("Name is ");
+        System.out.println("random ID is " + person1.getIDRandom());
+        System.out.println("Name is " + person1.getName());
         System.out.println("Surname is ");
         System.out.println("Status is ");
-        System.out.println("ID is ");
+        System.out.println("ID is " + person1.getID() + " Next ID is " + person1.getIDRandom() );
 
 
     }
